// src/Presentation/Mintada.Web/src/api/models/CoinTypeDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CoinTypeDto = {
    id?: number | string;
    issuerId?: number | string;
    title?: string;
    subtitle?: string | null;
    edgeImage?: string | null;
    period?: string | null;
    rarityIndex?: number | string | null;
    coinTypeSlug?: string;
    dateTimeInserted?: string;
    issueTypeId?: number | string;
    obverseImage?: string | null;
    reverseImage?: string | null;
};